+++ packages/cli/src/handleFatalError.ts
@@ -1,0 +1,22 @@
+import { logger } from './logger';
+import { spinner } from './spinner';
+
+export function handleFatalError(e: Error): any {
+  const errorText = `Unable to start GraphQL-Mesh: ${e.message}`;
+  if (spinner.isSpinning) {
+    spinner.fail(errorText);
+  } else {
+    logger.error(errorText);
+  }
+  if (process.env.DEBUG) {
+    logger.debug(
+      JSON.stringify({
+        ...e,
+        name: e.name,
+        stack: e.stack,
+        message: e.message,
+      })
+    );
+  }
+  process.exit(1);
+}