+++ packages/handlers/json-schema/src/json-schema-visitor.ts
@@ -138,16 +138,16 @@ export class JSONSchemaVisitor<TContext> {
     }
     if ('definitions' in def) {
       for (const propertyName in def.definitions) {
-        const definition: any = def.definitions[propertyName];
-        definition.name = definition.name || propertyName;
-        this.visit({ def: definition, propertyName, prefix, cwd, typeName: propertyName });
+        const definition = def.definitions[propertyName];
+        const result = this.visit({ def: definition, propertyName, prefix, cwd, typeName: propertyName });
+        this.visitedRefNameMap.set('#/definitions/' + propertyName, result);
       }
     }
     if ('$defs' in def) {
       for (const propertyName in def.$defs) {
-        const definition: any = def.$defs[propertyName];
-        definition.name = definition.name || propertyName;
-        this.visit({ def: definition, propertyName, prefix, cwd, typeName: propertyName });
+        const definition = def.$defs[propertyName];
+        const result = this.visit({ def: definition, propertyName, prefix, cwd, typeName: propertyName });
+        this.visitedRefNameMap.set('#/$defs/' + propertyName, result);
       }
     }
     let result: string;
@@ -197,7 +197,10 @@ export class JSONSchemaVisitor<TContext> {
         }
     }
     if ('oneOf' in def) {
-      result = this.visitOneOfReference({ oneOfReference: def, propertyName, prefix, cwd, typeName });
+      result = this.visitOneOfReference({ def, propertyName, prefix, cwd, typeName });
+    }
+    if ('const' in def) {
+      result = this.visitConst({ def });
     }
     if (!result && !ignoreResult) {
       console.warn(`Unknown JSON Schema definition for (${typeName || prefix}, ${propertyName})`);
@@ -453,32 +456,46 @@ export class JSONSchemaVisitor<TContext> {
     return 'Void';
   }
 
+  visitConst({ def }: { def: { const: string } }) {
+    const constValue = def.const;
+    const scalarName = constValue + 'ConstValue';
+    this.schemaComposer.add(new RegularExpression(scalarName, new RegExp(constValue)));
+    return scalarName;
+  }
+
   visitOneOfReference({
-    oneOfReference,
+    def,
     propertyName,
     prefix,
     cwd,
     typeName,
   }: {
-    oneOfReference: JSONSchemaOneOfDefinition;
+    def: JSONSchemaOneOfDefinition;
     propertyName: string;
     prefix: string;
     cwd: string;
     typeName?: string;
   }) {
-    let unionIdentifier = oneOfReference.title;
+    let unionIdentifier = def.title;
     if (!unionIdentifier) {
       unionIdentifier = prefix + '_' + propertyName;
     }
     const unionName = typeName || this.createName({ ref: unionIdentifier, cwd });
-    const types = oneOfReference.oneOf.map(def => this.visit({ def, propertyName, prefix, cwd }));
-
+    const types = def.oneOf.map(oneOfDef => this.visit({ def: oneOfDef, propertyName, prefix, cwd }));
     this.schemaComposer.createUnionTC({
       name: unionName,
       types,
       resolveType: (root: any) => {
         if (root.__typename) {
           return root.__typename;
+        }
+        const discriminatorPropertyName = def.discriminator?.propertyName;
+        if (discriminatorPropertyName) {
+          const discriminatorPropertyValue = root[discriminatorPropertyName];
+          if (discriminatorPropertyValue) {
+            const discriminatorTypeRef = def.discriminator.mapping[discriminatorPropertyValue];
+            return this.visitedRefNameMap.get(discriminatorTypeRef);
+          }
         }
         for (const typeName of types) {
           const typeDef = this.schemaComposer.getAnyTC(typeName);