+++ examples/grpc-example/CHANGELOG.md
@@ -1,5 +1,53 @@
 # grpc-example
 
+## 0.5.70
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/cli@0.23.0
+  - @graphql-mesh/grpc@0.8.11
+  - @graphql-mesh/transform-naming-convention@0.6.26
+
+## 0.5.69
+
+### Patch Changes
+
+- @graphql-mesh/grpc@0.8.10
+- @graphql-mesh/transform-naming-convention@0.6.25
+- @graphql-mesh/cli@0.22.1
+
+## 0.5.68
+
+### Patch Changes
+
+- Updated dependencies [970c43e2]
+- Updated dependencies [2cfa421a]
+  - @graphql-mesh/cli@0.22.0
+
+## 0.5.67
+
+### Patch Changes
+
+- Updated dependencies [ea3a183b]
+  - @graphql-mesh/cli@0.21.0
+
+## 0.5.66
+
+### Patch Changes
+
+- @graphql-mesh/grpc@0.8.9
+- @graphql-mesh/transform-naming-convention@0.6.24
+- @graphql-mesh/cli@0.20.2
+
+## 0.5.65
+
+### Patch Changes
+
+- @graphql-mesh/grpc@0.8.8
+- @graphql-mesh/transform-naming-convention@0.6.23
+- @graphql-mesh/cli@0.20.1
+
 ## 0.5.64
 
 ### Patch Changes