+++ examples/mongoose-example/CHANGELOG.md
@@ -1,5 +1,50 @@
 # mongoose-example
 
+## 0.5.68
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/cli@0.23.0
+  - @graphql-mesh/mongoose@0.6.43
+
+## 0.5.67
+
+### Patch Changes
+
+- @graphql-mesh/mongoose@0.6.42
+- @graphql-mesh/cli@0.22.1
+
+## 0.5.66
+
+### Patch Changes
+
+- Updated dependencies [970c43e2]
+- Updated dependencies [2cfa421a]
+  - @graphql-mesh/cli@0.22.0
+
+## 0.5.65
+
+### Patch Changes
+
+- Updated dependencies [ea3a183b]
+  - @graphql-mesh/cli@0.21.0
+
+## 0.5.64
+
+### Patch Changes
+
+- @graphql-mesh/mongoose@0.6.41
+- @graphql-mesh/cli@0.20.2
+
+## 0.5.63
+
+### Patch Changes
+
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/mongoose@0.6.40
+  - @graphql-mesh/cli@0.20.1
+
 ## 0.5.62
 
 ### Patch Changes