+++ examples/openapi-stripe/CHANGELOG.md
@@ -1,5 +1,57 @@
 # openapi-stripe
 
+## 0.5.71
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/cli@0.23.0
+  - @graphql-mesh/openapi@0.12.12
+
+## 0.5.70
+
+### Patch Changes
+
+- @graphql-mesh/openapi@0.12.11
+- @graphql-mesh/cli@0.22.1
+
+## 0.5.69
+
+### Patch Changes
+
+- Updated dependencies [970c43e2]
+- Updated dependencies [2cfa421a]
+  - @graphql-mesh/cli@0.22.0
+
+## 0.5.68
+
+### Patch Changes
+
+- Updated dependencies [ea3a183b]
+  - @graphql-mesh/cli@0.21.0
+
+## 0.5.67
+
+### Patch Changes
+
+- Updated dependencies [9354aeba]
+  - @graphql-mesh/openapi@0.12.10
+
+## 0.5.66
+
+### Patch Changes
+
+- Updated dependencies [938cca26]
+  - @graphql-mesh/openapi@0.12.9
+  - @graphql-mesh/cli@0.20.2
+
+## 0.5.65
+
+### Patch Changes
+
+- @graphql-mesh/openapi@0.12.8
+- @graphql-mesh/cli@0.20.1
+
 ## 0.5.64
 
 ### Patch Changes