+++ examples/soap-country-info/CHANGELOG.md
@@ -1,5 +1,54 @@
 # country-info-example
 
+## 0.5.66
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/cli@0.23.0
+  - @graphql-mesh/soap@0.5.3
+  - @graphql-mesh/transform-snapshot@0.7.16
+
+## 0.5.65
+
+### Patch Changes
+
+- @graphql-mesh/soap@0.5.2
+- @graphql-mesh/transform-snapshot@0.7.15
+- @graphql-mesh/cli@0.22.1
+
+## 0.5.64
+
+### Patch Changes
+
+- Updated dependencies [970c43e2]
+- Updated dependencies [2cfa421a]
+  - @graphql-mesh/cli@0.22.0
+
+## 0.5.63
+
+### Patch Changes
+
+- Updated dependencies [ea3a183b]
+  - @graphql-mesh/cli@0.21.0
+
+## 0.5.62
+
+### Patch Changes
+
+- @graphql-mesh/soap@0.5.1
+- @graphql-mesh/transform-snapshot@0.7.14
+- @graphql-mesh/cli@0.20.2
+
+## 0.5.61
+
+### Patch Changes
+
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/soap@0.5.0
+  - @graphql-mesh/transform-snapshot@0.7.13
+  - @graphql-mesh/cli@0.20.1
+
 ## 0.5.60
 
 ### Patch Changes