+++ examples/sqlite-chinook/CHANGELOG.md
@@ -1,5 +1,49 @@
 # chinook
 
+## 0.5.66
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/cli@0.23.0
+  - @graphql-mesh/tuql@0.4.43
+
+## 0.5.65
+
+### Patch Changes
+
+- @graphql-mesh/tuql@0.4.42
+- @graphql-mesh/cli@0.22.1
+
+## 0.5.64
+
+### Patch Changes
+
+- Updated dependencies [970c43e2]
+- Updated dependencies [2cfa421a]
+  - @graphql-mesh/cli@0.22.0
+
+## 0.5.63
+
+### Patch Changes
+
+- Updated dependencies [ea3a183b]
+  - @graphql-mesh/cli@0.21.0
+
+## 0.5.62
+
+### Patch Changes
+
+- @graphql-mesh/tuql@0.4.41
+- @graphql-mesh/cli@0.20.2
+
+## 0.5.61
+
+### Patch Changes
+
+- @graphql-mesh/tuql@0.4.40
+- @graphql-mesh/cli@0.20.1
+
 ## 0.5.60
 
 ### Patch Changes