+++ packages/handlers/json-schema/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @graphql-mesh/json-schema
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/types@0.31.0
+
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [48f38a4a]
+  - @graphql-mesh/types@0.30.1
+
+## 0.10.0
+
+### Minor Changes
+
+- 938cca26: feat(json-schema): allow custom error property with errorMessageField
+
+### Patch Changes
+
+- Updated dependencies [938cca26]
+  - @graphql-mesh/types@0.30.0
+
+## 0.9.8
+
+### Patch Changes
+
+- 1d21d02a: fix(fhir): handle fhir schema correctly
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/types@0.29.4
+
 ## 0.9.7
 
 ### Patch Changes