+++ packages/handlers/mongoose/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @graphql-mesh/mongoose
 
+## 0.6.43
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/types@0.31.0
+
+## 0.6.42
+
+### Patch Changes
+
+- Updated dependencies [48f38a4a]
+  - @graphql-mesh/types@0.30.1
+
+## 0.6.41
+
+### Patch Changes
+
+- Updated dependencies [938cca26]
+  - @graphql-mesh/types@0.30.0
+
+## 0.6.40
+
+### Patch Changes
+
+- 8ef29de1: fix(mongoose): throw errors if model is not valid
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/types@0.29.4
+
 ## 0.6.39
 
 ### Patch Changes