+++ packages/handlers/openapi/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @graphql-mesh/openapi
 
+## 0.12.12
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/types@0.31.0
+
+## 0.12.11
+
+### Patch Changes
+
+- Updated dependencies [48f38a4a]
+  - @graphql-mesh/types@0.30.1
+
+## 0.12.10
+
+### Patch Changes
+
+- 9354aeba: fix(openapi): remove misleading baseUrl warning
+
+## 0.12.9
+
+### Patch Changes
+
+- 938cca26: fix(openapi): fix handling default success code and no more get baseUrl from context
+- Updated dependencies [938cca26]
+  - @graphql-mesh/types@0.30.0
+
+## 0.12.8
+
+### Patch Changes
+
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/types@0.29.4
+
 ## 0.12.7
 
 ### Patch Changes