+++ packages/handlers/soap/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @graphql-mesh/soap
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [77327988]
+  - @graphql-mesh/types@0.31.0
+
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [48f38a4a]
+  - @graphql-mesh/types@0.30.1
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [938cca26]
+  - @graphql-mesh/types@0.30.0
+
+## 0.5.0
+
+### Minor Changes
+
+- 8ef29de1: feat(soap): ability to provide custom headers
+
+### Patch Changes
+
+- Updated dependencies [8ef29de1]
+  - @graphql-mesh/types@0.29.4
+
 ## 0.4.45
 
 ### Patch Changes